move namespace and introduce memory based index file size due to heavy use on swapping indexes from and to disk later on

+++ lib/io/chunk_collection_index_persistent.cpp
@@ -8,72 +8,189 @@ namespace uh::io
 
 // ---------------------------------------------------------------------
 
+
+namespace
+{
+
+// ---------------------------------------------------------------------
+
+std::filesystem::path index_path(std::unique_ptr<io::file>& collection_file)
+{
+    return collection_file->path().replace_extension(".index");
+}
+
+// ---------------------------------------------------------------------
+
+std::vector<std::pair<serialization::fragment_serialize_size_format, std::streamoff>>
+maybe_index_persist_chunk_collection(std::unique_ptr<io::file>& collection_file)
+{
+    auto filename_index = index_path(collection_file);
+    bool is_index_persisted = std::filesystem::exists(filename_index);
+
+    std::vector<std::pair<serialization::fragment_serialize_size_format, std::streamoff>> output_index;
+    std::streamoff collection_offset{};
+
+    if (is_index_persisted)
+    {
+        io::file read_index_file(filename_index, std::ios_base::binary | std::ios_base::in);
+
+        std::string read_buffer;
+        read_buffer.resize(sizeof(serialization::fragment_serialize_size_format));
+
+        while (io::read(read_index_file, read_buffer))
+        {
+            serialization::fragment_serialize_size_format skip_format;
+            std::istringstream ss(read_buffer);
+            skip_format.deserialize(ss);
+
+            output_index.emplace_back(skip_format, collection_offset);
+            collection_offset += skip_format.header_size + skip_format.content_size;
+        }
+
+        if (collection_offset < collection_file->size())
+        {
+            read_index_file.close();
+            std::filesystem::remove(read_index_file.path());
+            output_index.clear();
+        }
+        else
+            return output_index;
+    }
+
+    auto temporarily_cached_fragment_on_seekable_device =
+        io::fragment_on_seekable_device(*collection_file);
+
+    io::file write_index_file(filename_index, std::ios_base::binary | std::ios_base::out);
+
+    if (collection_file->size())
+    {
+        uint16_t index_entry_count{};
+        serialization::fragment_serialize_size_format skip_format;
+
+        do
+        {
+            skip_format = temporarily_cached_fragment_on_seekable_device.skip();
+
+            if (skip_format.content_size == 0)
+                break;
+
+            if (index_entry_count > std::numeric_limits<unsigned char>::max() + 1)
+            THROW(util::exception,
+                  "Indexing of chunk collection " + collection_file->path().string() + " exceeded limits");
+
+            output_index.emplace_back(skip_format, collection_offset);
+
+            io::write(write_index_file, skip_format.serialize().str());
+
+            collection_offset += skip_format.header_size + skip_format.content_size;
+            index_entry_count++;
+        }
+        while (skip_format.content_size > 0);
+    }
+
+    return output_index;
+}
+
+// ---------------------------------------------------------------------
+
+} // namespace
+
+// ---------------------------------------------------------------------
+
 chunk_collection_index_persistent::chunk_collection_index_persistent(std::unique_ptr<io::file>& chunk_collection_file)
     :
     std::vector<std::pair<serialization::fragment_serialize_size_format, std::streamoff>>(
         maybe_index_persist_chunk_collection(chunk_collection_file)),
     m_index_file(index_path(chunk_collection_file), std::ios_base::app),
-    m_workfile(chunk_collection_file)
+    m_workfile(chunk_collection_file),
+    m_index_file_size(m_index_file.size())
 {}
 
 // ---------------------------------------------------------------------
 
 void chunk_collection_index_persistent::erase_index_items(const std::vector<uint8_t>& at)
 {
-    if(forgotten){
-        maybe_index_persist_chunk_collection(m_workfile);
-        m_index_file = io::file(index_path(m_workfile), std::ios_base::app);
-        forgotten = false;
-    }
-    else{
-        auto seek_order_at_list = filtered_at_list_in_seek_order(at);
-        auto to_remove_it = begin();
+    maybe_recreate_index_file();
 
-        std::vector<uint16_t> delete_pos_list;
+    auto seek_order_at_list = filtered_at_list_in_seek_order(at);
+    auto to_remove_it = begin();
 
-        for (const auto at_item : seek_order_at_list)
+    std::vector<uint16_t> delete_pos_list;
+
+    for (const auto at_item : seek_order_at_list)
+    {
+        to_remove_it = find_address(at_item, to_remove_it);
+        auto remove_size_struct = to_remove_it->first;
+
+        std::for_each(to_remove_it, end(), [&remove_size_struct](
+            std::pair<serialization::fragment_serialize_size_format, std::streamoff>& index_pair)
         {
-            to_remove_it = find_address(at_item, to_remove_it);
-            auto remove_size_struct = to_remove_it->first;
-
-            std::for_each(to_remove_it, end(), [&remove_size_struct](
-                std::pair<serialization::fragment_serialize_size_format, std::streamoff>& index_pair)
-            {
-                index_pair.second -= (remove_size_struct.header_size + remove_size_struct.content_size);
-            });
-
-            delete_pos_list.push_back(to_remove_it->first.index_num);
-        }
-
-        io::temp_file erase_tmp(m_index_file.path().parent_path(), std::ios_base::out);
-        auto beg = begin();
-
-        std::string read_buffer;
-        read_buffer.resize(sizeof(serialization::fragment_serialize_size_format));
-
-        while (io::read(m_index_file, read_buffer))
-        {
-            if (std::none_of(delete_pos_list.cbegin(), delete_pos_list.cend(), [&beg](auto item)
-            {
-                return beg->first.index_num == item;
-            }))
-                io::write(erase_tmp, read_buffer);
-            beg++;
-        }
-
-        erase_tmp.close();
-        m_index_file.close();
-        erase_tmp.release_to(erase_tmp.path());
-        erase_tmp.rename(m_index_file.path());
-
-        std::erase_if(*this,[&delete_pos_list](auto& item){
-            return std::any_of(delete_pos_list.begin(),delete_pos_list.end(),[&item](const auto item2){
-                return item.first.index_num == item2;
-            });
+            index_pair.second -= (remove_size_struct.header_size + remove_size_struct.content_size);
         });
 
-        m_index_file = io::file(m_index_file.path(), std::ios_base::app);
+        delete_pos_list.push_back(to_remove_it->first.index_num);
     }
+
+    io::temp_file erase_tmp(m_index_file.path().parent_path(), std::ios_base::out);
+    auto beg = begin();
+
+    std::string read_buffer;
+    read_buffer.resize(sizeof(serialization::fragment_serialize_size_format));
+
+    while (io::read(m_index_file, read_buffer))
+    {
+        if (std::none_of(delete_pos_list.cbegin(), delete_pos_list.cend(), [&beg](auto item)
+        {
+            return beg->first.index_num == item;
+        }))
+            io::write(erase_tmp, read_buffer);
+        beg++;
+    }
+
+    erase_tmp.close();
+    m_index_file.close();
+    erase_tmp.release_to(erase_tmp.path());
+    erase_tmp.rename(m_index_file.path());
+
+    std::size_t deleted_index_file_size{};
+
+    std::erase_if(*this,
+                  [&delete_pos_list, &deleted_index_file_size](std::pair<serialization::fragment_serialize_size_format,
+                                                                         std::streamoff>& item)
+                  {
+                      bool to_be_deleted = std::any_of(delete_pos_list.begin(),
+                                                       delete_pos_list.end(),
+                                                       [&item](const auto item2)
+                                                       {
+                                                           return item.first.index_num == item2;
+                                                       });
+
+                      if (to_be_deleted)
+                      {
+                          deleted_index_file_size += item.first.header_size + item.first.content_size;
+                      }
+
+                      return to_be_deleted;
+                  });
+
+    m_index_file_size -= deleted_index_file_size;
+
+    m_index_file = io::file(m_index_file.path(), std::ios_base::app);
+}
+
+// ---------------------------------------------------------------------
+
+std::pair<serialization::fragment_serialize_size_format,
+          std::streamoff> chunk_collection_index_persistent::emplace_back_index(serialization::fragment_serialize_size_format write_format,
+                                                                                std::size_t emplace_size)
+{
+    auto tmp = emplace_back(write_format, emplace_size);
+
+    maybe_recreate_index_file();
+
+    io::write(m_index_file, back().first.serialize().str());
+
+    return tmp;
 }
 
 // ---------------------------------------------------------------------
@@ -155,11 +272,6 @@ uint16_t chunk_collection_index_persistent::free() const
 
 // ---------------------------------------------------------------------
 
-std::filesystem::path chunk_collection_index_persistent::getPath()
-{
-    return m_index_file.path();
-}
-
 // ---------------------------------------------------------------------
 
 uint8_t chunk_collection_index_persistent::next_free_address()
@@ -211,7 +323,7 @@ std::vector<uint8_t> chunk_collection_index_persistent::filtered_at_list_in_seek
 std::vector<std::pair<serialization::fragment_serialize_size_format, std::streamoff>>::iterator
 chunk_collection_index_persistent::find_address(uint8_t at,
                                                 std::vector<std::pair<serialization::fragment_serialize_size_format,
-                                                                       std::streamoff>>::iterator start_pos)
+                                                                      std::streamoff>>::iterator start_pos)
 {
     auto fragment_pos_element = std::find_if(start_pos, this->end(),
                                              [&at](const std::pair<
@@ -231,11 +343,33 @@ chunk_collection_index_persistent::find_address(uint8_t at,
 
 // ---------------------------------------------------------------------
 
-void chunk_collection_index_persistent::forget()
+void chunk_collection_index_persistent::maybe_forget_index_file()
 {
-    m_index_file.close();
-    std::filesystem::remove(m_index_file.path());
-    forgotten = true;
+    if (not m_index_file_forgotten)
+    {
+        m_index_file.close();
+        std::filesystem::remove(m_index_file.path());
+        m_index_file_forgotten = true;
+    }
+}
+
+// ---------------------------------------------------------------------
+
+void chunk_collection_index_persistent::maybe_recreate_index_file()
+{
+    if (m_index_file_forgotten)
+    {
+        maybe_index_persist_chunk_collection(m_workfile);
+        m_index_file = io::file(index_path(m_workfile), std::ios_base::app);
+        m_index_file_forgotten = false;
+    }
+}
+
+// ---------------------------------------------------------------------
+
+std::size_t chunk_collection_index_persistent::get_index_file_size() const
+{
+    return m_index_file_size;
 }
 
 // ---------------------------------------------------------------------
